Changed to use temporary files everywhere - but haven't yet made sure that they will be deleted when finished with.
Added in statistics code from previous DunesGIS project, converting to work
with ArcGIS 10.0 (gp -> arcpy, uncapitalising first letter of methods etc).

At the moment it just prints out a CSV file line, it can be converted later on to run as a function which can be called for each file given as input.